input gatherelements matmul gemm
