+++ src/acetone_nnet/generator/__init__.py
@@ -57,10 +57,15 @@ from .layers import (
     Flatten,
     FlattenDefault,
     Gather,
+GatherDefault,
     GatherElements,
+GatherElementsDefault,
     Gemm,
+GemmDefault,
     InputLayer,
+InputLayerDefault,
     MatMul,
+MatMulDefault,
     Maximum,
     MaxPooling2D,
     Minimum,
