+++ src/acetone_nnet/generator/layers/matmul/MatMul.py
@@ -131,45 +131,6 @@ class MatMul(Layer):
 
     def generate_inference_code_layer(self: Self) -> str:
         """Generate computation code for layer."""
-        mustach_hash = {}
-
-        mustach_hash["name"] = self.name
-        mustach_hash["idx"] = f"{self.idx:02d}"
-        mustach_hash["comment"] = self.activation_function.comment
-        mustach_hash["road"] = self.path
-        mustach_hash["size"] = self.size
-
-        if self.activation_function.name != "linear":
-            mustach_hash["non_linear"] = True
-            mustach_hash["activation_function"] = self.activation_function.write_activation_str(
-                f"tensor_temp[j + {self.output_width}*(i + {self.output_height}*f)]")
-
-        mustach_hash["shared_dimension"] = self.shared_dimension
-        mustach_hash["output_channels"] = self.output_channels
-        mustach_hash["output_height"] = self.output_height
-        mustach_hash["output_width"] = self.output_width
-
-        if self.side == 0:
-            mustach_hash["output_str_left"] = self.previous_layer[0].output_str
-            mustach_hash["output_str_right"] = f"weights_{self.name}_{self.idx:02d}"
-        elif self.side == 1:
-            mustach_hash["output_str_right"] = self.previous_layer[0].output_str
-            mustach_hash["output_str_left"] = f"weights_{self.name}_{self.idx:02d}"
-        elif self.side == 2:
-            mustach_hash["output_str_left"] = self.previous_layer[0].output_str
-            mustach_hash["output_str_right"] = self.previous_layer[1].output_str
-
-        if self.fused_layer:
-            mustach_hash["fused_layer"] = self.fused_layer.write_activation_str(
-                self.local_var,
-                self.idx,
-                "i")
-
-        with open(self.template_path / "layers" / "template_MatMul.c.tpl") as template_file:
-            template = template_file.read()
-        template_file.close()
-
-        return pystache.render(template, mustach_hash)
 
     def forward_path_layer(
             self: Self,
