+++ tests/tests_inference/tests_layer/test_c_reference/test_GatherElements.py
@@ -30,7 +30,7 @@ class TestGatherElements(acetoneTestCase.AcetoneTestCase):
     def testGatherElements1(self):
         in_shape = list(np.random.randint(1, 100, size=3))
         in_shape = (1, int(in_shape[0]), int(in_shape[1]), int(in_shape[2]))
-        out_shape = (1, int(np.random.randint(1, in_shape[1])), int(in_shape[2]), int(in_shape[3]))
+        out_shape = (1, int(np.random.randint(1, high=in_shape[1])), int(in_shape[2]), int(in_shape[3]))
 
         model_input_name = "X"
         X = onnx.helper.make_tensor_value_info(model_input_name,
@@ -42,7 +42,7 @@ class TestGatherElements(acetoneTestCase.AcetoneTestCase):
                                                out_shape)
 
         indices_name = "indice"
-        indices = indices = np.random.randint(in_shape[1], size=out_shape)
+        indices = np.random.randint(0, high=in_shape[1], size=out_shape)
         indices_initializer = acetoneTestCase.create_initializer_tensor(name=indices_name,
                                                                         tensor_array=indices,
                                                                         data_type=onnx.TensorProto.INT32)
@@ -90,7 +90,7 @@ class TestGatherElements(acetoneTestCase.AcetoneTestCase):
                                                out_shape)
 
         indices_name = "indice"
-        indices = indices = np.random.randint(in_shape[2], size=out_shape)
+        indices = np.random.randint(in_shape[2], size=out_shape)
         indices_initializer = acetoneTestCase.create_initializer_tensor(name=indices_name,
                                                                         tensor_array=indices,
                                                                         data_type=onnx.TensorProto.INT32)
@@ -138,7 +138,7 @@ class TestGatherElements(acetoneTestCase.AcetoneTestCase):
                                                out_shape)
 
         indices_name = "indice"
-        indices = indices = np.random.randint(in_shape[3], size=out_shape)
+        indices = np.random.randint(in_shape[3], size=out_shape)
         indices_initializer = acetoneTestCase.create_initializer_tensor(name=indices_name,
                                                                         tensor_array=indices,
                                                                         data_type=onnx.TensorProto.INT32)
